+++ R/app_ui.R
@@ -40,16 +40,32 @@ app_ui <- function(request) {
                 shinydashboard::tabItems(
                     shinydashboard::tabItem(
                         tabName = "home",
-                        mod_home_ui()
+                        mod_home_ui("home")
                     ), # end of tabItem 1
                     shinydashboard::tabItem(
                         tabName = "parameters",
                         shiny::h1("Parameters"),
                         mod_data_import_ui("data_import"),
+                        shiny::fluidRow(
+                            shiny::column(width = 6,
+                                shinydashboard::box(
+                                    status = "warning",
+                                    width = 12,
+                                    solidHeader = TRUE,
+                                    title = shiny::h3("Please choose a Project name"),
+                                    shiny::textInput(inputId = "project_title",
+                                                     label = NULL,
+                                                     value = "",
+                                                     placeholder = "my project title")
+                                )
+
+                            )
+
+                        ),
                         mod_plate_dimensions_ui("p_dim"),
-                        mod_special_wells_ui("special1"),
+                        mod_special_wells_ui("special1"), # forbidden wells
                         mod_plate_specifications_ui("plate"),
-                        mod_special_wells_ui("special2"),
+                        mod_special_wells_ui("special2"), # not randomized wells
                         shiny::fluidRow(
                             shiny::column(
                                 width = 6,
@@ -57,7 +73,7 @@ app_ui <- function(request) {
                                     status = "warning",
                                     width = 12,
                                     solidHeader = TRUE,
-                                    title = shiny::h3("6 - Number of iterations"),
+                                    title = shiny::h3("Number of iterations"),
                                     shiny::h4(
                                     "Please specify the maximum number of iterations
                                     that WPM can perform. Default value is 20."),
@@ -87,13 +103,11 @@ app_ui <- function(request) {
                             )# end of column
                         ) # end of fluiRow 3
                     ),# end of tabItem 2
-                    #
                     shinydashboard::tabItem(
                         tabName = "results",
                         shiny::h1("Your results"),
                         mod_data_export_ui("data_export")
                     )# end of tabItem 3
-
                 ) # end of tabItems
             ),
             skin = "yellow"))