+++ R/fct_imports.R
@@ -1,18 +1,16 @@
-##' Convert a CSV File into a dataframe
+##' Convert a CSV File into a valid dataframe for WPM
 ##'
 ##' @description This function converts a CSV into a dataframe to make it
 ##' usable by the shiny application of wpm as well as by the wrapper function
-##' (version of wpm in command line)
+##' (version of wpm in command line). Be sure that the first column of the CSV 
+##' file corresponds to samples names.
 ##'
 ##' @param dt_path file path.
-##' @param head logical value, indicates if the file has a header on it's first
-##' line.
-##' @param qt character, the set of quoting characters. To disable quoting
-##' altogether, use quote = "".
-##' @param sep character, the field separator character. Values on each line
-##' of the file are separated by this character.
-##' @param row_names logical value, does the file have rownames?
-##'
+##' @param row_names logical value, indicates if the file has rownames or not.
+##' @param gp_field the column name indicating the grouping factor for the 
+##' samples in the csv. If there is no grouping factor, then gp_field must be 
+##' set to NULL or "none".
+##' @param ... parameters to give to read.csv2 function
 ##' @return a dataframe containing 3 fields: Sample, Group and ID.
 ##'
 ##' @examples
@@ -20,53 +18,79 @@
 ##'                    "Group" = c("A","A","B", "C"))
 ##' tf <- tempfile()
 ##' write.csv2(test, tf, row.names = FALSE)
-##' convertCSV(tf, head = TRUE, sep = ";")
+##' convertCSV(tf, gp_field = "Group", header = TRUE, sep = ";")
 ##'
 ##' # if there are row names in the CSV file
 ##' write.csv2(test, tf)
-##' convertCSV(tf, head = TRUE, sep = ";", row_names = TRUE)
+##' convertCSV(tf, row_names = TRUE, gp_field="Group", header = TRUE, sep = ";")
+##' 
+##' # if there is no grouping factor in the CSV file
+##' convertCSV(tf, row_names = TRUE, gp_field ="none", header = TRUE, sep = ";")
+##' # gives the same output as the previous example
+##' convertCSV(tf, row_names = TRUE, header = TRUE, sep = ";")
 ##' @export
-convertCSV <- function(dt_path, head = FALSE, qt = "", sep = "",
-                       row_names = FALSE){
+convertCSV <- function(dt_path, row_names = FALSE, gp_field = NULL, ...){
     if (row_names) {
-        df <- utils::read.csv2(dt_path,
-                               header = head,
-                               quote = qt,
-                               sep = sep,
-                               stringsAsFactors = FALSE,
-                               row.names = 1
-        )
+        df <- utils::read.csv2(dt_path, row.names = 1, ...)
     }else{
-        df <- utils::read.csv2(dt_path,
-                               header = head,
-                               quote = qt,
-                               sep = sep,
-                               stringsAsFactors = FALSE
-        )
+        df <- utils::read.csv2(dt_path, ...)
     }
-    ## check if file contains groups or not
-    if (length(colnames(df)) == 1) {
-        colnames(df) <- "Sample"
-        df$Group <- as.factor(1)
+    
+    if (is.null(gp_field)) {
+        df <- data.frame(Sample = df[,1], Group = as.factor(1))
+    }else if (gp_field == "none") {
+        df <- data.frame(Sample = df[,1], Group = as.factor(1))
     }else{
-        colnames(df) <- c("Sample", "Group")
+        # check if user enter an existing field
+        if (gp_field %in% colnames(df)) {
+            df <- data.frame(Sample = df[,1],
+                             Group = as.factor(df[[gp_field]]))
+        }else{
+            stop("The group field must be an existing column name for the 
+                phenotype Data", call. = FALSE)
+        }
     }
     df$Sample <- as.character(df$Sample)
     ## a unique ID for each sample which will be used by the drawMap function
     df$ID <- seq_len(nrow(df))
+    return(df)
+}
 
+##' Reshape a phenotype dataframe for WPM
+##' @param pD_df a dataframe containing the phenotype data
+##' @param gp_field the column name indicating the grouping factor for the 
+##' samples
+##' @return a dataframe containing 3 fields: Sample, Group and ID.
+##'
+##' @noRd
+reshapeDataframe <- function(pD_df, gp_field){
+    if (is.null(gp_field)) {
+        df <- data.frame(Sample = rownames(pD_df), Group = as.factor(1))
+    }else{
+        # check if user enter an existing field
+        if (gp_field %in% colnames(pD_df)) {
+            df <- data.frame(Sample = rownames(pD_df),
+                             Group = as.factor(pD_df[[gp_field]]))
+        }else{
+            stop("The group field must be an existing column name for the 
+                phenotype Data", call. = FALSE)
+        }
+    }
+    df$Sample <- as.character(df$Sample)
+    df$ID <- seq_len(nrow(df))
     return(df)
 }
 
 
-##' Convert an ExpressionSet or MsnSet into a dataframe
+
+##' Convert the phenotype data of an ExpressionSet or MsnSet into a dataframe 
+##' for WPM
 ##'
 ##' @description This function converts an ExpressionSet/MsnSet object into a
 ##' dataframe to make it usable by the shiny application of wpm as well as by
 ##' the wrapper function (version of wpm in command line)
 ##'
-##' @param eSet_obj an ExpressionSet/MsnSet object contaning at least the
-##' phenotype Data
+##' @param eSet_obj an ExpressionSet/MsnSet object contaning the phenotype data
 ##' @param gp_field character, corresponding to the phenotype data used to
 ##' categorize samples into distinct groups if any
 ##'
@@ -85,17 +109,30 @@ convertCSV <- function(dt_path, head = FALSE, qt = "", sep = "",
 ##'
 ##' @export
 convertESet <- function(eSet_obj, gp_field = NULL){
-
     pD_df <- Biobase::pData(eSet_obj)
+    df <- reshapeDataframe(pD_df, gp_field)
+    return(df)
+}
 
-    if (is.null(gp_field)) {
-        df <- data.frame(Sample = rownames(pD_df), Group = as.factor(1))
-    }else{
-        df <- data.frame(Sample = rownames(pD_df),
-                         Group = as.factor(pD_df[[gp_field]]))
-    }
 
-    df$ID <- seq_len(nrow(df))
 
+##' Convert the phenotype data of a SummarizedExperiment into a dataframe for WPM
+##'
+##' @param se_object a SummarizedExperiment object containing the phenotype data
+##' @param gp_field character, corresponding to the phenotype data used to
+##' categorize samples into distinct groups if any
+##' @return a dataframe containing 3 fields: Sample, Group and ID.
+##' @examples 
+##' nrows <- 200
+##' ncols <- 6
+##' counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
+##' colData <- data.frame(Treatment=rep(c("ChIP", "Input"), 3),
+##'                     row.names=LETTERS[1:6])
+##' se <- SummarizedExperiment::SummarizedExperiment(assays=list(counts=counts), colData=colData)
+##' convertSE(se, "Treatment")
+##' @export
+convertSE <- function(se_object, gp_field = NULL){
+    pD_df <- SummarizedExperiment::colData(se_object)
+    df <- reshapeDataframe(pD_df, gp_field)
     return(df)
 }